chore(tts): decommission int4 for TTS models (defaults → bf16/int8)
int4 quantization degrades TTS quality too much, so it's being removed as a
shipped option for the TTS models. Repoint every int4 TTS default to the best
available non-int4 variant — bf16 where it exists, int8 where it doesn't:

  - Qwen3-TTS: default model id → 1.7B-bf16 (the production model). The 0.6B
    `.base` variant → 0.6B-8bit (no 0.6B bf16 bundle exists); drop the redundant
    `.base8bit` case. Rename the int4 `large4bit` Talker/CodePredictor configs to
    a precision-neutral `large` base (bf16 floor); the config ladder now respects
    any `bits` so the loader can still build a 4-bit config.
  - CosyVoice3: default → 0.5B-bf16; drop the `4bit` CLI variant.
  - Magpie-TTS: drop the `.int4` enum case (no bf16 bundle); int8 is the floor.
    CoreML weight extraction + tokenizer assets now source the 8-bit MLX bundle.
  - VoxCPM2: already defaults to bf16; drop the selectable `int4` CLI variant.
  - AudioServer back-compat shims + ModelRegistry entries → the new defaults.
  - Tests updated to the surviving variants; the int4-specific tests removed.

The one exception is Qwen3-TTS CustomVoice, which ships only as int4 — no 8bit
or bf16 bundle exists yet. It stays on int4 behind a tracked TODO, and the
loader keeps the ability to construct a 4-bit config for it.

+++ Sources/MagpieTTS/Downloader.swift
@@ -15,7 +15,7 @@ public enum MagpieTTSDownloader {
     }
 
     public static func ensureDownloaded(
-        variant: MagpieTTSVariant = .int4,
+        variant: MagpieTTSVariant = .int8,
         progressHandler: ((Double) -> Void)? = nil
     ) async throws -> Paths {
         let repoId = variant.huggingFaceRepoId
